service-workers/service-worker/import-module-scripts.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Tests for module import: ServiceWorker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script type="text/javascript" src="/workers/modules/resources/import-test-cases.js"></script>
<body>
<script type="text/javascript">
  function import_test(testCase) {
    promise_test(async () => {
      // TODO(asamidoi): ASK this test fails when 13 and 14 lines move to after
      // registration.installing.onmessage line.
      const registration = await navigator.serviceWorker.register(testCase.scriptURL, {type: "module"});
      registration.installing.postMessage("Send message for tests from main script.");
      navigator.serviceWorker.onmessage = msgEvent => {
        assert_array_equals(msgEvent.data, testCase.expectation);
      };
    }, testCase.description);
  }

  testCases.forEach(import_test);
</script>
</body>
